backport patches switch pac4j version from rc to ga
