+++ core/cas-server-core-util-api/src/test/java/org/apereo/cas/util/crypto/GlibcCryptPasswordEncoderTests.java
@@ -4,7 +4,7 @@ import lombok.extern.slf4j.Slf4j;
 import lombok.val;
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 /**
  * This tests {@link GlibcCryptPasswordEncoder}.
@@ -16,6 +16,25 @@ import static org.junit.jupiter.api.Assertions.*;
 public class GlibcCryptPasswordEncoderTests {
 
     private static final String PASSWORD_CLEAR = "12345abcDEF!$";
+
+    private static boolean testEncodingRoundtrip(final String algorithm) {
+        val encoder = new GlibcCryptPasswordEncoder(algorithm, 0, null);
+
+        val passwordHash = encoder.encode(PASSWORD_CLEAR);
+        LOGGER.debug("Password [{}] was encoded by algorithm [{}] to hash [{}]", PASSWORD_CLEAR, algorithm, passwordHash);
+
+        val match = encoder.matches(PASSWORD_CLEAR, passwordHash);
+        LOGGER.debug("Does password [{}] match original password [{}]: [{}]", passwordHash, PASSWORD_CLEAR, match);
+
+        return match;
+    }
+
+    private static boolean testMatchWithDifferentSalt(final String algorithm, final String encodedPassword) {
+        val encoder = new GlibcCryptPasswordEncoder(algorithm, 0, null);
+        val match = encoder.matches(PASSWORD_CLEAR, encodedPassword);
+        LOGGER.debug("Does password [{}] match original password [{}]: [{}]", encodedPassword, PASSWORD_CLEAR, match);
+        return match;
+    }
 
     @Test
     public void sha512EncodingTest() {
@@ -44,24 +63,4 @@ public class GlibcCryptPasswordEncoderTests {
         assertTrue(testEncodingRoundtrip("42xyz"));
         assertTrue(testMatchWithDifferentSalt("aB", "aB4fMcNOggJoQ"));
     }
-
-    private static boolean testEncodingRoundtrip(final String algorithm) {
-        val encoder = new GlibcCryptPasswordEncoder(algorithm, 0, null);
-
-        val passwordHash = encoder.encode(PASSWORD_CLEAR);
-        LOGGER.debug("Password [{}] was encoded by algorithm [{}] to hash [{}]", PASSWORD_CLEAR, algorithm, passwordHash);
-
-        val match = encoder.matches(PASSWORD_CLEAR, passwordHash);
-        LOGGER.debug("Does password [{}] match original password [{}]: [{}]", passwordHash, PASSWORD_CLEAR, match);
-
-        return match;
-    }
-
-    private static boolean testMatchWithDifferentSalt(final String algorithm, final String encodedPassword) {
-        val encoder = new GlibcCryptPasswordEncoder(algorithm, 0, null);
-        val match = encoder.matches(PASSWORD_CLEAR, encodedPassword);
-        LOGGER.debug("Does password [{}] match original password [{}]: [{}]", encodedPassword, PASSWORD_CLEAR, match);
-        return match;
-    }
-
 }
