+++ core/cas-server-core-web-api/src/main/java/org/apereo/cas/web/view/RestfulUrlTemplateResolver.java
@@ -7,7 +7,6 @@ import org.apereo.cas.web.support.WebUtils;
 
 import lombok.extern.slf4j.Slf4j;
 import lombok.val;
-
 import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.http.HttpResponse;
