+++ core/cas-server-core-web/src/test/java/org/apereo/cas/AllTestsSuite.java
@@ -3,7 +3,6 @@ package org.apereo.cas;
 
 import org.apereo.cas.web.RegisteredServiceResponseHeadersEnforcementFilterTests;
 import org.apereo.cas.web.SimpleUrlValidatorFactoryBeanTests;
-
 import org.apereo.cas.web.WebjarValidationTests;
 import org.junit.platform.runner.JUnitPlatform;
 import org.junit.platform.suite.api.SelectClasses;
