+++ core/cas-server-core-webflow-mfa-api/src/main/java/org/apereo/cas/web/flow/actions/MultifactorAuthenticationBypassAction.java
@@ -5,7 +5,6 @@ import org.apereo.cas.web.support.WebUtils;
 
 import lombok.extern.slf4j.Slf4j;
 import lombok.val;
-
 import org.springframework.webflow.execution.Event;
 import org.springframework.webflow.execution.RequestContext;
