+++ gradle.properties
@@ -37,7 +37,7 @@ targetCompatibility=11
 # in order to fully override component versions that are provided by Spring Boot.
 ##################################################################################
 
-lombokVersion=1.18.10
+lombokVersion=1.18.12
 aspectjVersion=1.9.4
 errorProneVersion=2.3.2
 errorproneJavacVersion=9+181-r4173-1
@@ -197,8 +197,8 @@ jodaTimeVersion=2.10.5
 reflectionsVersion=0.9.11
 guavaVersion=28.1-jre
 
-pac4jSpringWebmvcVersion=4.0.0-RC2
-pac4jVersion=4.0.0-RC2
+pac4jSpringWebmvcVersion=4.0.0
+pac4jVersion=4.0.0
 
 statsdVersion=3.1.0
 micrometerVersion=1.3.0
