+++ support/cas-server-support-duo-core-mfa/src/main/java/org/apereo/cas/adaptors/duo/authn/DefaultDuoSecurityMultifactorAuthenticationProvider.java
@@ -12,7 +12,6 @@ import lombok.NonNull;
 import lombok.Setter;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
-
 import org.springframework.cloud.context.config.annotation.RefreshScope;
 
 /**
