+++ support/cas-server-support-duo/src/main/java/org/apereo/cas/adaptors/duo/config/DuoSecurityConfiguration.java
@@ -16,7 +16,6 @@ import org.apereo.cas.web.flow.resolver.CasWebflowEventResolver;
 import org.apereo.cas.web.flow.resolver.impl.CasWebflowEventResolutionConfigurationContext;
 
 import lombok.extern.slf4j.Slf4j;
-
 import lombok.val;
 import org.springframework.beans.factory.ObjectProvider;
 import org.springframework.beans.factory.annotation.Autowired;
