+++ support/cas-server-support-pac4j-webflow/src/test/java/org/apereo/cas/web/flow/DelegatedClientAuthenticationActionTests.java
@@ -46,9 +46,9 @@ import org.junit.jupiter.api.Test;
 import org.pac4j.core.client.BaseClient;
 import org.pac4j.core.client.Clients;
 import org.pac4j.core.context.JEEContext;
-import org.pac4j.core.context.Pac4jConstants;
 import org.pac4j.core.context.WebContext;
 import org.pac4j.core.context.session.JEESessionStore;
+import org.pac4j.core.util.Pac4jConstants;
 import org.pac4j.oauth.client.FacebookClient;
 import org.pac4j.oauth.client.TwitterClient;
 import org.pac4j.oauth.credentials.OAuth20Credentials;
